requirements.txt
django==4.1.4
djangorestframework==3.14.0 
